rename : constants 폴더안에 있는 filterList 파일명변경, filterList 안에 변수명 스네이크기법으로 모두 변경(#7)
