+++ src/constants/filterList.js
@@ -6,7 +6,7 @@ const FILTER_LABELS = {
   latest: "최신순",
 };
 
-const categoryOptionList = [
+const CATEGORYOPTIONLIST = [
   { value: "daily", name: "🏠 일상" },
   { value: "hobby", name: "🎨 취미/여가" },
   { value: "travel", name: "✈️ 여행" },
@@ -15,7 +15,7 @@ const categoryOptionList = [
   { value: "other", name: "🌀 기타" },
 ];
 
-const priorityOptionList = [
+const PRIORITYOPTIONLIST = [
   {
     value: "urgent",
     name: "긴급",
@@ -35,12 +35,12 @@ const priorityOptionList = [
   },
 ];
 
-const shareOption = [
+const SHAREOPTION = [
   { value: "personalSchedule", name: "개인일정", icon: <FaUser /> },
   { value: "sharedSchedule", name: "공유하기", icon: <FaShareAlt /> },
 ];
 
-const repeatOptionList = [
+const REPEATOPTIONLIST = [
   { value: "none", name: "반복없음" },
   { value: "daily", name: "매일반복" },
   { value: "weekly", name: "매주반복" },
@@ -48,15 +48,15 @@ const repeatOptionList = [
   { value: "yearly", name: "매년반복" },
 ];
 
-const latestOptionList = [
+const LATESTOPTIONLIST = [
   { value: "newest", name: "최신순" },
   { value: "oldest", name: "오래된순" },
 ];
 
 const FILTERS = {
-  category: { label: "카테고리", options: categoryOptionList },
-  priority: { label: "중요도", options: priorityOptionList },
-  share: { label: "공유", options: shareOption },
-  repeat: { label: "반복", options: repeatOptionList },
-  latest: { label: "최신순", options: latestOptionList },
+  category: { label: "카테고리", options: CATEGORYOPTIONLIST },
+  priority: { label: "중요도", options: PRIORITYOPTIONLIST },
+  share: { label: "공유", options: SHAREOPTION },
+  repeat: { label: "반복", options: REPEATOPTIONLIST },
+  latest: { label: "최신순", options: LATESTOPTIONLIST },
 };
